api/requirements.txt
tensorflow==2.11.1
Streamlit==1.22.0
python-multipart
pillow
matplotlib
numpys